implement save bill detail

+++ app/src/main/java/id/fahrizal/billreminder/data/model/BillDetail.kt
@@ -9,5 +9,5 @@ data class BillDetail(
     val message: String,
     val notifDate: Long,
     val isPaid :Boolean = false,
-    @PrimaryKey val id: Int? = null
+    @PrimaryKey val id: Long? = null
 )
